setup main file
