+++ README.md
@@ -18,7 +18,7 @@ pip3 install matchscraper
 
 ## Usage
 ```shell
-$ matchscraper --match cbv 1623
+$ matchscraper --fed cbv --match 1623
 ```
 
 ```
