+++ matchscraper/settings.py
@@ -90,10 +90,3 @@ ROBOTSTXT_OBEY = True
 # Set settings whose default value is deprecated to a future-proof value
 REQUEST_FINGERPRINTER_IMPLEMENTATION = '2.7'
 TWISTED_REACTOR = 'twisted.internet.asyncioreactor.AsyncioSelectorReactor'
-
-FEEDS = {
-    'data/%(name)s.csv': {
-        'format': 'csv',
-        'overwrite': True
-    }
-}
